+++ src/main/java/step5/racingcar3/domain/RacingGame.java
@@ -19,15 +19,4 @@ public class RacingGame {
         // Do nothing
     }
 
-    public CarList winners() {
-        Position maxPosition = maxPosition();
-        return new CarList(
-            cars.value().stream().filter(car -> car.position().equals(maxPosition)).collect(Collectors.toList()));
-    }
-
-    private Position maxPosition() {
-        return Collections.max(cars.value().stream().map(Car::position).collect(Collectors.toList()));
-    }
-
-
 }
